{{cookiecutter.project_slug}}/requirements.txt
arrow==1.2.3
asgiref==3.6.0
backports.zoneinfo==0.2.1;python_version<"3.9"
binaryornot==0.4.4
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cookiecutter==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.53.0
djangorestframework==3.14.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
uritemplate==4.1.1
urllib3==1.26.15
celery==4.4.1
redis==3.4.1
django-environ==0.10.0